+++ src/site/markdown/release_notes.md
@@ -11,6 +11,8 @@ Version TBD has the following notable changes:
 
 * Morphlines Library
     * Made xquery and xslt commands also compatible with woodstox-3.2.7 (not just woodstox-4.x).
+    * Added morphline command to read from RCFile.
+        * [readRCFile](cdk-morphlines/morphlinesReferenceGuide.html#readRCFile)
 
 ## Version 0.8.0